Initial Commit
Node,.js http protocol, no express for, just http, request, path.

If you throw up words, it should give you back an array of well formed JSON objects, the object is composed of the "word", which also the call to the JSON object in the map we create.  If you dont want to make and parse a webpage to make a  an object , you can just use scrape to scrape the html.  

If you want to then get the words out of that html, css and live js, use domparse above.  

If you want to see what the return data looks like, check out the files called "output" and dictionary,js  

If you want a test set of the words we to place into the grammar test maker and pos/meaning parser--JS objects containing all the required keys to handle each passage are in well structured JSON objects.  The value of each property['key'](now)=>WORD.value, or, just the word.  

Lastly if you want to see what I passed my first scraper ob--- see examplescraper,.jss


To be most efficient, we should likely static the standard properties of the JSON objects/